+++ packages/app/studio/src/ui/Footer.tsx
@@ -9,6 +9,7 @@ import {FooterLabel} from "@/service/FooterLabel"
 import {ProjectMeta, StudioPreferences} from "@opendaw/studio-core"
 import {Colors} from "@opendaw/studio-enums"
 import {UserCounter} from "@/UserCounter"
+import {AudioData} from "@opendaw/lib-dsp"
 
 const className = Html.adoptStyleSheet(css, "footer")
 
@@ -83,6 +84,21 @@ export const Footer = ({lifecycle, service}: Construct) => {
                          }, "footer-show-fps-meter"))
                      }}>0
             </article>
+            <article title="Samples (GC)"
+                     onInit={element => {
+                         const lifeSpan = lifecycle.own(new Terminator())
+                         lifecycle.own(StudioPreferences.catchupAndSubscribe(show => {
+                             element.classList.toggle("hidden", !show)
+                             if (show) {
+                                 lifeSpan.own(Runtime.scheduleInterval(() => {
+                                     element.textContent = AudioData.count().toString()
+                                 }, 1000))
+                             } else {
+                                 lifeSpan.terminate()
+                             }
+                         }, "footer-show-samples-memory"))
+                     }}>0
+            </article>
             <div style={{display: "contents"}}
                  onInit={element => {
                      const lifeSpan = lifecycle.own(new Terminator())
@@ -108,11 +124,9 @@ export const Footer = ({lifecycle, service}: Construct) => {
                      }, "footer-show-build-infos"))
                  }}/>
             <article title="Users"
-                     onInit={async element => {
+                     onInit={element => {
                          const counter = new UserCounter("https://api.opendaw.studio/users/user-counter.php")
-                         element.textContent = String(await counter.start())
-                         setInterval(async () => element.textContent = String(await counter.updateUserCount()), 30000)
-                         window.addEventListener("beforeunload", () => counter.stop())
+                         counter.subscribe(count => element.textContent = String(count))
                      }}>#
             </article>
             <div style={{display: "contents"}}