+++ packages/studio/adapters/src/sample/SampleLoaderManager.ts
@@ -1,4 +1,4 @@
-import {UUID} from "@opendaw/lib-std"
+import {Terminable, UUID} from "@opendaw/lib-std"
 import {SampleLoader} from "./SampleLoader"
 
 export interface SampleLoaderManager {
@@ -6,4 +6,5 @@ export interface SampleLoaderManager {
     record(loader: SampleLoader): void
     invalidate(uuid: UUID.Bytes): void
     remove(uuid: UUID.Bytes): void
+    register(uuid: UUID.Bytes): Terminable
 }