+++ packages/studio/core-processors/src/SampleManagerWorklet.ts
@@ -3,7 +3,7 @@ import {EngineToClient, SampleLoader, SampleLoaderManager, SampleLoaderState} fr
 import {Observer, Option, SortedSet, Subscription, Terminable, UUID} from "@opendaw/lib-std"
 import {AudioData} from "@opendaw/lib-dsp"
 
-class AudioLoaderWorklet implements SampleLoader {
+class SampleLoaderWorklet implements SampleLoader, Terminable {
     readonly peaks: Option<Peaks> = Option.None
     readonly #state: SampleLoaderState = {type: "idle"}
 
@@ -18,25 +18,29 @@ class AudioLoaderWorklet implements SampleLoader {
 
     subscribe(_observer: Observer<SampleLoaderState>): Subscription {return Terminable.Empty}
     invalidate(): void {}
+    terminate(): void {this.#data = Option.None}
 
-    toString(): string {return `{AudioLoaderWorklet}`}
+    toString(): string {return `{SampleLoaderWorklet}`}
 }
 
-export class SampleManagerWorklet implements SampleLoaderManager {
+export class SampleManagerWorklet implements SampleLoaderManager, Terminable {
     readonly #engineToClient: EngineToClient
-    readonly #set: SortedSet<UUID.Bytes, SampleLoader>
+    readonly #set: SortedSet<UUID.Bytes, SampleLoaderWorklet>
 
     constructor(engineToClient: EngineToClient) {
         this.#engineToClient = engineToClient
-        this.#set = UUID.newSet<SampleLoader>(handler => handler.uuid)
+        this.#set = UUID.newSet<SampleLoaderWorklet>(({uuid}) => uuid)
     }
 
     record(_loader: SampleLoader): void {}
-
     getOrCreate(uuid: UUID.Bytes): SampleLoader {
-        return this.#set.getOrCreate(uuid, uuid => new AudioLoaderWorklet(uuid, this.#engineToClient))
+        return this.#set.getOrCreate(uuid, uuid => new SampleLoaderWorklet(uuid, this.#engineToClient))
     }
-
-    remove(_uuid: UUID.Bytes) {}
-    invalidate(_uuid: UUID.Bytes) {}
+    remove(uuid: UUID.Bytes): void {this.#set.removeByKey(uuid).terminate()}
+    register(_uuid: UUID.Bytes): Terminable {return Terminable.Empty}
+    invalidate(_uuid: UUID.Bytes): void {}
+    terminate(): void {
+        this.#set.forEach(loader => loader.terminate())
+        this.#set.clear()
+    }
 }