+++ packages/studio/core-processors/src/devices/instruments/NanoDeviceProcessor.ts
@@ -105,6 +105,11 @@ export class NanoDeviceProcessor extends AudioProcessor implements InstrumentDev
         this.#peakBroadcaster.process(this.#audioOutput.getChannel(0), this.#audioOutput.getChannel(1))
     }
 
+    terminate(): void {
+        super.terminate()
+        this.loader = Option.None
+    }
+
     toString(): string {return `{NanoDevice}`}
 }